requirements.txt
Flask==3.0.0
langchain==0.1.0
langgraph==0.0.20
ollama==0.1.6
python-dotenv==1.0.0
playwright==1.48.0